requirements.txt
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-mapbox-location-field==1.4.0
django-rest-auth==0.9.5
django-sslserver==0.22
django-storages==1.9.1
djangorestframework==3.11.0
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
idna==2.8
Markdown==3.1.1
oauthlib==3.1.0
Pillow==7.0.0
pip==9.0.1
postgres==3.0.0
protobuf==3.11.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
setuptools==45.2.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
wheel==0.34.2
